Disable dark mode support for now.
The dark mode support has been disabled for now, as it was acheived
by applying same set of tailwind classes to a lot of components. I guess
it would be better to use @apply directive and compose classes for
common use cases like light and dark typography, background and border
etc. Therefore I am disabling it for now and will do a rewrite of it
using tailwind's @apply directive.